// src/modules/futures/dto/future-order-info-response.dto.ts
export class FutureOrderInfoResponseDto {
  symbol: string;
  orderId: string;
  price: string;
  size: string;
  orderType: string;
  side: string;
  status: string;
  priceAvg: string;
  margin: string;
  baseVolume: string;
  quoteVolume: string;
  feeCurrency: string;
  fee: string;
  cTime: string;
  uTime: string;
}