// src/modules/futures/dto/future-position-info-response.dto.ts
export class FuturePositionInfoResponseDto {
  symbol: string;
  side: string;
  size: string;
  orderId: string;
  marginMode: string;
  marginCurrency: string;
  marginBalance: string;
  marginRatio: string;
  marginAvailable: string;
  marginLocked: string;
  leverage: string;
  priceAvg: string;
  feeCurrency: string;
  fee: string;
  cTime: string;
  uTime: string;
}