// src/modules/futures/dto/future-position-history-info-response.dto.ts
export class FuturePositionHistoryInfoResponseDto {
  symbol: string;
  openTime: string;
  closeTime: string;
  marginMode: string;
  avgEntryPrice: string;
  avgClosePrice: string;
  side: string;
  size: string;
  orderId: string;
  marginCurrency: string;
  marginBalance: string;
  marginRatio: string;
  leverage: string;
  unrealizedPnl: string;
  feeCurrency: string;
  fee: string;
}